refactor: アイデア出し終了画面リファクタリング #175
