+++ front/src/features/end-session/components/ResultSection/ResultSection.module.scss
@@ -1,0 +1,10 @@
+.resultWrapper {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 16px;
+  width: 100%;
+  margin-top: 16px;
+  margin-bottom: 40px;
+}
