+++ front/src/features/end-session/components/ResultThisMonth/ResultThisMonth.tsx
@@ -1,0 +1,16 @@
+import styles from "@/features/end-session/components/ResultThisMonth/ResultThisMonth.module.scss";
+
+export default function ResultThisMonth({
+  ideaMemosCountThisMonth,
+}: {
+  ideaMemosCountThisMonth: number;
+}) {
+  return (
+    <div className={styles.resultContainer}>
+      <div className={styles.resultTitle}>今月出したアイデアの総数</div>
+      <div className={styles.resultCount}>
+        <span>{ideaMemosCountThisMonth}</span> 個
+      </div>
+    </div>
+  );
+}
