+++ front/src/features/end-session/components/index.ts
@@ -1,0 +1,6 @@
+import CompareComment from "@/features/end-session/components/CompareComment/CompareComment";
+import ResultSection from "@/features/end-session/components/ResultSection/ResultSection";
+import ResultThisMonth from "@/features/end-session/components/ResultThisMonth/ResultThisMonth";
+import ResultThisSession from "@/features/end-session/components/ResultThisSession/ResultThisSession";
+
+export { CompareComment, ResultSection, ResultThisMonth, ResultThisSession };
